requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.25.0
datasets==2.14.4
deepspeed==0.10.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.1.3
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
grpcio==1.57.0
hjson==3.1.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lit==15.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
ninja==1.11.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
peft==0.5.0
protobuf==4.24.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.0
pydantic==1.10.12
Pygments==2.16.1
python-dateutil==2.8.2
python-rapidjson==1.10
pytz==2023.3
PyYAML==6.0.1
ray==2.6.3
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.3
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.11.1
tabulate==0.9.0
tokenizers==0.13.3
torch==2.0.1+cu118
torchtyping==0.1.4
tqdm==4.66.1
transformers==4.32.0
triton==2.0.0
tritonclient==2.36.0
typeguard==4.1.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.8
xxhash==3.3.0
yarl==1.9.2
